added new hook – useFiltering
